+++ packages/littoral-templates-node/src/index.ts
@@ -1,21 +1,4 @@
-/**
- * EOL-related constants and functions.
- */
-
-/**
- * Common EOL styles.
- */
-export const eolStyles = {
-    lf: "\n",
-    crlf: "\r\n"
-} as const
-
-
-/**
- * The value of the EOL string: typically, either "\n" (=LF-style, compatible with POSIX/Linux/macOS), or "\r\n" (=CRLF-style, compatible with Windows).
- */
-export let eol: string = eolStyles.lf
-
+import {setEOLExplicitly} from "littoral-templates"
 
 /**
  * Sets the EOL to the OS-dependent value.
@@ -27,7 +10,7 @@ export const setEOLStyleFromOS = async () => {
     await import("os")
         .then((osPackage) => {
             if (osPackage) {
-                eol = osPackage.EOL
+                setEOLExplicitly(osPackage.EOL)
             } else {
                 printWarningOnConsole()
             }
@@ -41,23 +24,4 @@ export const setEOLStyleFromOS = async () => {
             }
         })
 }
-
-
-/**
- * Sets the EOL explicitly to the given value.
- * (This is especially useful for testing.)
- */
-export const setEOLExplicitly = (newEol: string) => {
-    eol = newEol
-}
-
-
-const splitRegex = new RegExp(/\r*\n/)
-
-/**
- * Splits the given string on EOLs (of any style) into an array of strings,
- * each corresponding to one line.
- */
-export const splitOnEOL = (str: string): string[] =>
-    str.split(splitRegex)